[#286] moved layouting-related code into OrganicLayouter
