+++ src/test/java/de/bonndan/nivio/output/layout/OrganicLayouterTest.java
@@ -1,6 +1,7 @@
 package de.bonndan.nivio.output.layout;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import de.bonndan.nivio.assessment.Assessment;
 import de.bonndan.nivio.input.FileFetcher;
 import de.bonndan.nivio.input.Indexer;
 import de.bonndan.nivio.input.InputFormatHandlerFactory;
@@ -11,11 +12,16 @@ import de.bonndan.nivio.input.dto.ItemDescription;
 import de.bonndan.nivio.input.dto.LandscapeDescription;
 import de.bonndan.nivio.input.dto.RelationDescription;
 import de.bonndan.nivio.input.http.HttpService;
+import de.bonndan.nivio.model.Item;
 import de.bonndan.nivio.model.Landscape;
+import de.bonndan.nivio.model.LandscapeFactory;
 import de.bonndan.nivio.output.RenderingTest;
 import de.bonndan.nivio.output.icons.IconService;
 import de.bonndan.nivio.output.icons.LocalIcons;
 import de.bonndan.nivio.output.icons.ExternalIcons;
+import de.bonndan.nivio.output.map.svg.MapStyleSheetFactory;
+import de.bonndan.nivio.output.map.svg.SVGDocument;
+import de.bonndan.nivio.output.map.svg.SVGRenderer;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.context.ApplicationEventPublisher;
@@ -26,7 +32,11 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 
+import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.when;
 
 class OrganicLayouterTest extends RenderingTest {
 
@@ -170,5 +180,25 @@ class OrganicLayouterTest extends RenderingTest {
         indexer = new Indexer(landscapeRepository, formatFactory, linkHandlerFactory, mock(ApplicationEventPublisher.class),  iconService);
 
         debugRender("/src/test/resources/example/example_csv", false);
+    }
+
+    @Test
+    public void shiftGroupsAndItems() {
+
+        //given
+        String path = "/src/test/resources/example/inout";
+        Landscape landscape = getLandscape(path + ".yml");
+
+
+        //when
+        OrganicLayouter layouter = new OrganicLayouter();
+        LayoutedComponent lc = layouter.layout(landscape);
+
+        LayoutedComponent itemComponent = lc.getChildren().get(0).getChildren().get(0);
+        assertNotNull(itemComponent);
+
+        //check items are shifted
+        assertEquals(691.4090488048637, itemComponent.getX()); //margin + group offset + own offset
+        assertEquals(1459.8723644530933, itemComponent.getY()); //margin + group offset + own offset
     }
 }
